+++ static/dtale/MeasureText.jsx
@@ -1,4 +1,5 @@
 import $ from "jquery";
+import _ from "lodash";
 import React from "react";
 
 function measureText(str) {
@@ -10,7 +11,7 @@ function measureText(str) {
       "white-space": "nowrap",
       visibility: "hidden",
     });
-  return o.width() + 20; // 5px padding on each side
+  return _.round(o.width()) + 20; // 5px padding on each side
 }
 
 require("./MeasureText.css");